Maximize signature input window automatically
Was previously fixed to a smaller size - like the user input window was - but now it takes up the whole screen (besides the taskbar) when opened. This provides more room to draw the signature, which is more convenient for users/customers signing out their devices. The canvas and image have been resized to fit the larger window, but with room at the bottom for the submit button.